Add event service

+++ src/main/resources/db/migration/V1__create_identity_table.sql
@@ -2,17 +2,17 @@ CREATE SEQUENCE SEQ_IDENTITY_ID INCREMENT BY 1 MINVALUE 1 NO MAXVALUE START WITH
 
 CREATE TABLE T_IDENTITY (
   -- BaseEntity
-  ID            INTEGER                                                              NOT NULL,
-  OWNER_ID      INTEGER DEFAULT 0                                                    NOT NULL,
+  ID            BIGINT                                                               NOT NULL,
+  OWNER_ID      BIGINT DEFAULT 0                                                     NOT NULL,
   CREATE_TIME   TIMESTAMP DEFAULT CURRENT_TIMESTAMP                                  NOT NULL,
   UPDATE_TIME   TIMESTAMP                                                            NULL,
   IS_DELETED    CHAR(1) DEFAULT 'N'                                                  NOT NULL,
   --
   SALT          VARCHAR(16) DEFAULT '-'                                              NOT NULL,
   PASSWORD      VARCHAR(64) DEFAULT '-'                                              NOT NULL,
-  CREATE_BY     INTEGER                                                              NOT NULL,
+  CREATE_BY     BIGINT                                                               NOT NULL,
   TYPE          VARCHAR(64) DEFAULT 'USER'                                           NOT NULL,
-  VALID_SECONDS INTEGER                                                              NULL,
+  VALID_SECONDS BIGINT                                                               NULL,
   IS_BLOCKED    CHAR(1) DEFAULT 'N'                                                  NOT NULL,
   CONSTRAINT PK_IDENTITY PRIMARY KEY (ID),
   CONSTRAINT FK_IDENTITY_OWNER FOREIGN KEY (OWNER_ID) REFERENCES T_IDENTITY (ID),
